fix(decode): all decode bugs fixed, but unsure about API again
If I can make it faster by making inline not overwrite and also append,
I will.

Its kinda slower now that the bugs are fixed. Still fast though, but not
as crazy

Hopefully I can do something about that too. I only just fixed them, it
can probably be optimized.

Still a work in progress
