+++ Makefile
@@ -12,8 +12,6 @@ LIBFLAG      ?= -shared
 CFLAGS       += $(LIBFLAG) -I"$(LUA_INCDIR)"
 TESTDIR      := $(SRC)/tests
 SRCS         := $(SRC)/src/tomlua.c \
-                $(SRC)/src/decode_str.c \
-                $(SRC)/src/decode_inline_value.c \
                 $(SRC)/src/decode.c \
                 $(SRC)/src/encode.c \
                 $(SRC)/src/dates.c
